Example Data Set with Admitted Patients

An auto-generated data set containing fictitious patients admitted to hospitals.

Usage

admitted_patients

Format

A tibble/data.frame with 250 observations and 7 variables:

date
date of hospital admission

patient_id
ID of the patient (fictitious)

gender
gender of the patient

age
age of the patient

age_group
age group of the age of the patient, generated with AMR::age_groups()

hospital
ID of the hospital, from A to D

ward
type of ward, either ICU or Non-ICU