plot2() Options

These are the package options for usage in plot2(). These can all be set using options(), and their values can be retrieved using getOption().

Package Options

big_mark
Default character for a thousands separator

dec_mark
Default character for a decimal separator

plot2.colour
Default colour to use in plots

plot2.colour_background
Default background colour to use in plots

plot2.colour_font_axis
Default colour of the axis text (axis.title.x and axis.title.x)

plot2.colour_font_primary
Default colour of the plot title and plot tag

plot2.colour_font_secondary
Default colour of the plot subtitle

plot2.colour_sf
Default colour of the borders in sf plots

plot2.colour_sf_fill
Default colour of the area fills in sf plots

plot2.default_type
Default plot type

plot2.font
Default plot font family

plot2.silent
Default value to show console messages when plotting

plot2.text_factor
Default text factor to use in plots

plot2.theme
Default plot theme